chore: refactor ui components (#1172)
